backend/server/requirements.txt
django==3.2.5
djangorestframework==3.12.4
mysqlclient==2.0.3
django-cors-headers==3.7.0
requests==2.26.0